+++ lambda/handlers/index.js
@@ -1,0 +1,9 @@
+const newReminderHandler = require('./newReminderHandler')
+const quizHandler = require('./quizHandler')
+const medsTakenHandler = require('./medsTakenHandler')
+
+module.exports = { 
+    newReminderHandler, 
+    quizHandler,
+    medsTakenHandler
+}
