+++ lambda/index.js
@@ -3,6 +3,7 @@ const user_id = 1; // TODO: has to be an amazon ID
 const PERMISSIONS = ['alexa::alerts:reminders:skill:readwrite'];
 const utils = require('./utils/Utils')
 const API = require('./utils/apiUtils')
+const handlers = require('./handlers')
 const quizTime = '15:00' // TODO this needs to come from the backend at some point
 
 const LaunchRequestHandler = {
@@ -36,31 +37,22 @@ const LaunchRequestHandler = {
 };
 
 const QuizIntentHandler = {
-  canHandle(handlerInput) {
+  canHandle({ requestEnvelope }) {
     return (
-      Alexa.getRequestType(handlerInput.requestEnvelope) === "IntentRequest" &&
-      Alexa.getIntentName(handlerInput.requestEnvelope) === "QuizIntent"
+      Alexa.getRequestType(requestEnvelope) === "IntentRequest" &&
+      Alexa.getIntentName(requestEnvelope) === "QuizIntent"
     );
   },
   async handle(handlerInput) {
-    const { mood: {value: mood}, stiffness: {value: stiffness}, slowness: {value: slowness}, tremor: {value: tremor}} = handlerInput.requestEnvelope.request.intent.slots
-    const quizAnswers = { mood, stiffness, slowness, tremor };
-    const quiz = await API.postQuizAnswers(user_id, {...quizAnswers})
-    let speakOut = '';
-    const speakOut = (quiz && quiz.completed_at) ? 
-    'Your answers have been recorded'
-    :'Sorry, your answers could not be recorded. Please try again.'
-    return handlerInput.responseBuilder
-      .speak(speakOut)
-      .getResponse();
+    handlers.quizHandler(handlerInput)
   }
 };
 
 const PairDeviceIntentHandler = { // TODO: pull this into it's own file. It's huge! 
-  canHandle(handlerInput) {
+  canHandle({ requestEnvelope }) {
     return (
-      Alexa.getRequestType(handlerInput.requestEnvelope) === "IntentRequest" &&
-      Alexa.getIntentName(handlerInput.requestEnvelope) === "PairDeviceIntent"
+      Alexa.getRequestType(requestEnvelope) === "IntentRequest" &&
+      Alexa.getIntentName(requestEnvelope) === "PairDeviceIntent"
     );
   },
   async handle(handlerInput) {
@@ -76,79 +68,20 @@ const newReminderIntentHandler = {
       return Alexa.getRequestType(handlerInput.requestEnvelope) === 'IntentRequest'
           && Alexa.getIntentName(handlerInput.requestEnvelope) === 'newReminderIntent';
   },
-  async handle(handlerInput) {
-      const client = handlerInput.serviceClientFactory.getReminderManagementServiceClient();
-      const requestEnvelope = handlerInput.requestEnvelope;
-      const responseBuilder = handlerInput.responseBuilder;
-      const permissions = requestEnvelope.context.System.user.permissions
-      if (!permissions) {
-        // if no permissions, nag the user to grant them
-        return responseBuilder
-          .speak('I need permission to create reminders. Take a look at your alexa app to grant them.')
-          .withAskForPermissionsConsentCard(PERMISSIONS)
-          .getResponse()
-      }
-      switch (requestEnvelope.request.intent.confirmationStatus) {
-        case 'CONFIRMED':
-          // Alexa confirmed intent, so clear to create reminder
-          break;
-        case 'DENIED':
-          // Alexa disconfirmed the intent; not creating reminder
-          return responseBuilder
-            .speak(`Ok, I will not create any reminders`)
-            .reprompt('')
-            .getResponse();
-        default:
-          //delegate back to Alexa to get confirmation
-          return responseBuilder
-            .addDelegateDirective()
-            .getResponse();
-      }
-      try {
-        // TODO: delete reminders that are not required according to the schedule. Nuclear option?
-          const meds = await API.getMedicationList(user_id)
-          const { alerts: presentReminders } = await client.getReminders();
-          const filteredMeds = utils.filterMedsAgainstExistingReminders(meds, presentReminders)
-          const medsReminders = utils.reminderBuilder(filteredMeds)
-          medsReminders.forEach(async reminder => {
-            await client.createReminder(reminder).then(console.log)
-          })
-          const quiz = await utils.checkForQuizReminder(presentReminders, quizTime) // TODO: patch existing quiz reminder if time changed
-          if (!quiz) {
-            const quizReminder = utils.createQuizReminder(quizTime)
-            await client.createReminder(quizReminder).then(console.log)
-          }
-      } catch (error) {
-        if (error.name !== 'ServiceError') {
-          console.log(`error: ${error.stack}`);
-          return responseBuilder
-            .speak(`Something went wrong with the reminders`)
-            .getResponse();
-        }
-        throw error;
-      }
-    return responseBuilder
-      .speak('Reminders are up to date. Would you like to do anything else?')
-      .reprompt('Can I help you with anything else?')
-      .getResponse()
+  handle(handlerInput) {
+    handlers.newReminderHandler(handlerInput)
   }
 };
 
 const medsTakenHandler = {
-  canHandle(handlerInput) {
+  canHandle({ requestEnvelope }) {
     return (
-      Alexa.getRequestType(handlerInput.requestEnvelope) === "IntentRequest" &&
-      Alexa.getIntentName(handlerInput.requestEnvelope) === "medsTakenIntent"
+      Alexa.getRequestType(requestEnvelope) === "IntentRequest" &&
+      Alexa.getIntentName(requestEnvelope) === "medsTakenIntent"
     );
   },
-  async handle(handlerInput) {
-    const result = await API.postMedsTaken(user_id)
-    const speakOut = result ?
-    "Ok,  i've logged that you've taken your medication."
-    : "I'm sorry. I could not log that you've taken your medication."  
-    handlerInput.responseBuilder()
-      .speak(speakOut)
-      .getResponse()
+  handle(handlerInput) {
+    handlers.medsTakenHandler(handlerInput)
   }
 }
 
@@ -160,7 +93,7 @@ const HelpIntentHandler = {
     );
   },
   handle(handlerInput) {
-    const speakOutput = "I'm here to help you track of your medication and health. Try saying take a quiz.";
+    const speakOutput = "I'm here to help you keep track of your medication and health. Try saying take a quiz.";
     return handlerInput.responseBuilder
       .speak(speakOutput)
       .reprompt(speakOutput)
@@ -172,10 +105,8 @@ const CancelAndStopIntentHandler = {
   canHandle(handlerInput) {
     return (
       Alexa.getRequestType(handlerInput.requestEnvelope) === "IntentRequest" &&
-      (Alexa.getIntentName(handlerInput.requestEnvelope) ===
-        "AMAZON.CancelIntent" ||
-        Alexa.getIntentName(handlerInput.requestEnvelope) ===
-          "AMAZON.StopIntent")
+      (Alexa.getIntentName(handlerInput.requestEnvelope) === "AMAZON.CancelIntent" ||
+        Alexa.getIntentName(handlerInput.requestEnvelope) === "AMAZON.StopIntent")
     );
   },
   handle(handlerInput) {
@@ -238,5 +169,5 @@ exports.handler = Alexa.SkillBuilders.custom()
   )
   .withApiClient(new Alexa.DefaultApiClient())
   .addErrorHandlers(ErrorHandler)
-  .withCustomUserAgent(`Hayley_smells/v1`)
+  .withCustomUserAgent(`medirep-alexa/v1`)
   .lambda();
